requirements.txt
Flask==2.2.5
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.4.4
python-dotenv==1.0.0
pillow==10.0.0
werkzeug==2.3.7
